requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# LLM
google-genai>=1.0.0

# Agent Framework
langgraph>=0.2.0
langchain-core>=0.3.0

# Database
aiosqlite>=0.19.0

# Utilities
json-repair>=0.30.0
aiohttp>=3.9.0
websockets>=12.0
